@@ -1,61 +1,99 @@
 {
   "name": "broken-link-checker",
-  "description": "Find broken links, missing images, etc in your HTML.",
-  "version": "0.7.8",
+  "description": "Find broken links, missing images, etc within your HTML.",
+  "version": "0.8.0",
   "license": "MIT",
-  "author": "Steven Vachon <contact@svachon.com> (https://www.svachon.com/)",
-  "repository": "stevenvachon/broken-link-checker",
-  "main": "lib",
+  "author": "Steven Vachon <contact@svachon.com> (https://svachon.com)",
+  "repository": "github:stevenvachon/broken-link-checker",
+  "main": "lib-cjs",
   "bin": {
     "blc": "bin/blc",
     "broken-link-checker": "bin/blc"
   },
   "dependencies": {
-    "bhttp": "^1.2.1",
-    "calmcard": "~0.1.1",
-    "chalk": "^1.1.3",
-    "char-spinner": "^1.0.1",
-    "condense-whitespace": "^1.0.0",
+    "auto-tunnel": "github:stevenvachon/auto-tunnel",
+    "chalk": "^2.4.2",
+    "condense-whitespace": "^2.0.0",
+    "deep-freeze-node": "^1.1.3",
     "default-user-agent": "^1.0.0",
-    "errno": "~0.1.4",
-    "extend": "^3.0.0",
-    "humanize-duration": "^3.9.1",
-    "http-equiv-refresh": "^1.0.0",
-    "is-stream": "^1.0.1",
+    "errno": "~0.1.7",
+    "gauge": "^2.7.4",
+    "got": "^9.6.0",
+    "http-equiv-refresh": "^2.0.1",
+    "humanize-duration": "^3.21.0",
+    "is-stream": "^2.0.0",
     "is-string": "^1.0.4",
-    "limited-request-queue": "^2.0.0",
-    "link-types": "^1.1.0",
-    "maybe-callback": "^2.1.0",
-    "nopter": "~0.3.0",
-    "parse5": "^3.0.2",
-    "robot-directives": "~0.3.0",
-    "robots-txt-guard": "~0.1.0",
-    "robots-txt-parse": "~0.0.4",
-    "urlcache": "~0.7.0",
-    "urlobj": "0.0.11"
+    "isurl": "^4.0.1",
+    "keyscan": "^1.7.0",
+    "limited-request-queue": "^5.1.0",
+    "link-types": "^3.0.0",
+    "list-to-array": "^1.1.0",
+    "lodash": "^4.17.15",
+    "longest": "^2.0.1",
+    "matcher": "^2.0.0",
+    "node-notifier": "^6.0.0",
+    "optionator": "~0.8.2",
+    "parse-srcset": "^1.0.2",
+    "parse5": "^5.1.0",
+    "parse5-parser-stream": "^5.1.0",
+    "robot-directives": "github:stevenvachon/robot-directives",
+    "robots-txt-guard": "~0.2.1",
+    "robots-txt-parse": "^1.0.1",
+    "strip-ansi": "^5.2.0",
+    "supports-semigraphics": "^1.0.1",
+    "url-relation": "github:stevenvachon/url-relation",
+    "urlcache": "github:stevenvachon/urlcache",
+    "walk-parse5": "^2.0.0"
   },
   "devDependencies": {
-    "chai": "^3.5.0",
-    "chai-as-promised": "^6.0.0",
-    "chai-like": "~0.2.10",
+    "@babel/cli": "^7.6.2",
+    "@babel/core": "^7.6.2",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.4.4",
+    "@babel/plugin-proposal-numeric-separator": "^7.2.0",
+    "@babel/plugin-proposal-optional-catch-binding": "^7.2.0",
+    "@babel/plugin-proposal-optional-chaining": "^7.6.0",
+    "@babel/plugin-proposal-private-methods": "^7.6.0",
+    "@babel/preset-env": "^7.6.2",
+    "@babel/register": "^7.6.2",
+    "babel-eslint": "^10.0.3",
+    "babel-plugin-add-module-exports": "^1.0.2",
+    "basic-auth": "^2.0.1",
+    "basic-auth-header": "^1.0.1",
+    "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
+    "chai-subset": "github:stevenvachon/chai-subset",
     "chai-things": "~0.2.0",
-    "es6-promise": "^4.1.0",
-    "mocha": "^3.0.2",
-    "object.assign": "^4.0.4",
-    "slashes": "^1.0.5",
-    "st": "^1.2.0"
+    "coveralls": "^3.0.6",
+    "escape-string-regexp": "^2.0.0",
+    "eslint": "^6.4.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsdoc": "^15.9.5",
+    "eslint-plugin-sort-destructure-keys": "^1.3.3",
+    "eslint-plugin-you-dont-need-lodash-underscore": "^6.7.0",
+    "mocha": "^6.2.0",
+    "nock": "^11.3.5",
+    "nyc": "^14.1.1",
+    "void-elements": "^3.1.0"
   },
   "engines": {
-    "node": ">= 0.10"
+    "node": ">= 12"
   },
   "scripts": {
-    "test": "mocha test/ --reporter spec --check-leaks --bail",
-    "test-watch": "mocha test/ --reporter spec --check-leaks --bail -w"
+    "build": "babel lib/ --config-file=./scripts/babel.config.js --out-dir=lib-cjs/ --source-maps",
+    "ci": "npm run lint && npm test && nyc report --reporter=text-lcov | coveralls",
+    "lint": "npm run lint:cjs ; npm run lint:esm",
+    "lint:cjs": "eslint --config=scripts/eslintrc.cjs.js scripts/",
+    "lint:esm": "eslint --config=scripts/eslintrc.esm.js lib/ test/",
+    "posttest": "nyc report --reporter=text-summary --reporter=html",
+    "prepublishOnly": "npm run lint && npm test && npm run build",
+    "test": "nyc --exclude=scripts/ --silent mocha test/ --bail --check-leaks --require=scripts/register-babel",
+    "test:watch": "npm test -- --watch --watch-extensions=js,json"
   },
   "files": [
     "bin",
     "lib",
-    "license"
+    "lib-cjs"
   ],
   "keywords": [
     "404",
