@@ -1,510 +1,607 @@
-"use strict";
-var blc            = require("./");
-var defaultOptions = require("./internal/defaultOptions");
-var pkg            = require("../package.json");
+/* eslint-disable no-console */
+import {BROKEN_REASON, EXCLUDED_REASON, HTML_INDEX, HTTP_RESPONSE_WAS_CACHED, IS_BROKEN, ORIGINAL_URL, REBASED_URL, WAS_EXCLUDED} from "./internal/Link";
+import DEFAULT_OPTIONS from "./internal/defaultOptions";
+import {END_EVENT, ERROR_EVENT, HTML_EVENT, JUNK_EVENT, LINK_EVENT, PAGE_EVENT, QUEUE_EVENT, SITE_EVENT} from "./internal/events";
+import Gauge from "gauge";
+import {GET_METHOD} from "./internal/methods";
+import {gray, green, red, white, yellow} from "chalk";
+import {HtmlUrlChecker, SiteChecker} from "./";
+import humanizeDuration from "humanize-duration";
+import longest from "longest";
+import {make_scanner as scanKeys} from "keyscan";
+import notifier from "node-notifier";
+import stripAnsi from "strip-ansi";
+import supportsSemigraphics from "supports-semigraphics";
+import {themes as gaugeThemes} from "gauge/themes";
+import {version as packageVersion} from "../package.json";
 
-var chalk = require("chalk");
-var humanizeDuration = require("humanize-duration");
-var nopter = require("nopter");
-var spinner = require("char-spinner");
 
-process.on("unhandledRejection", function(reason, p) {
-	log( nopter.error.fatal("Unhandled Rejection", reason, "Error") );
-	process.exit(1);
-});
 
-function cli()
+const title = "Broken Link Checker";
+let checker,checkerOptions,gauge,keyScanner,logOptions,pauseMessage,spinner,stats,urls;
+
+
+
+const argsToOptions = args =>
 {
-	var filterLevel = "The types of tags and attributes that are considered links.\n";
-	filterLevel += "  0: clickable links\n";
-	filterLevel += "  1: 0 + media, iframes, meta refreshes\n";
-	filterLevel += "  2: 1 + stylesheets, scripts, forms\n";
-	filterLevel += "  3: 2 + metadata\n";
-	filterLevel += "  Default: "+defaultOptions.filterLevel;
-	
-	this.nopter = new nopter();
-	
-	this.nopter.config(
+	const renames =
 	{
-		title: "Broken Link Checker",
-		description: pkg.description,
-		version: pkg.version,
-		name: "blc",
-		options:
+		exclude: "excludedKeywords",
+		excludeExternal: "excludeExternalLinks",
+		excludeInternal: "excludeInternalLinks",
+		follow: "followRobotExclusions",
+		hostRequests: "maxSocketsPerHost",
+		include: "includedKeywords",
+		ordered: "maintainLinkOrder",
+		requests: "maxSockets"
+	};
+
+	return Object.entries(args).reduce((opts, [argName, argValue]) =>
+	{
+		if (argName in renames)
 		{
-			"exclude":
-			{
-				rename: "excludedKeywords",
-				info: "A keyword/glob to match links against. Can be used multiple times.",
-				type: [String, Array],
-				default: defaultOptions.excludedKeywords
-			},
-			"exclude-external":
-			{
-				rename: "excludeExternalLinks",
-				short: "e",
-				info: "Will not check external links.",
-				type: Boolean
-			},
-			"exclude-internal":
-			{
-				rename: "excludeInternalLinks",
-				short: "i",
-				info: "Will not check internal links.",
-				type: Boolean
-			},
-			"filter-level":
-			{
-				info: filterLevel,
-				type: Number,
-				default: defaultOptions.filterLevel
-			},
-			"follow":
-			{
-				rename: "followRobotExclusions",
-				short: "f",
-				info: "Force-follow robot exclusions.",
-				type: Boolean
-			},
-			"get":
-			{
-				short: "g",
-				info: "Change request method to GET.",
-				type: Boolean
-			},
-			"help":
-			{
-				short: ["h","?"],
-				info: "Display this help text.",
-				type: Boolean
-			},
-			"input":
-			{
-				info: "URL to an HTML document.",
-				type: require("url")
-			},
-			"host-requests":
-			{
-				rename: "maxSocketsPerHost",
-				info: "Concurrent requests limit per host.",
-				type: Number,
-				default: defaultOptions.maxSocketsPerHost
-			},
-			"ordered":
-			{
-				rename: "maintainLinkOrder",
-				short: "o",
-				info: "Maintain the order of links as they appear in their HTML document.",
-				type: Boolean
-			},
-			"recursive":
-			{
-				short: "r",
-				info: "Recursively scan (\"crawl\") the HTML document(s).",
-				type: Boolean
-			},
-			"requests":
-			{
-				rename: "maxSockets",
-				info: "Concurrent requests limit.",
-				type: Number,
-				default: defaultOptions.maxSockets
-			},
-			"user-agent":
-			{
-				info: "The user agent to use for link checks.",
-				type: String,
-				default: defaultOptions.userAgent
-			},
-			"verbose":
-			{
-				short: "v",
-				info: "Display excluded links.",
-				type: Boolean
-			},
-			"version":
-			{
-				short: "V",
-				info: "Display the app version.",
-				type: Boolean
-			}
-		},
-		aliases: ["input"]
-	});
-}
-
-
-
-cli.prototype.input = function(args, showArgs)
-{
-	//var testing = args !== undefined;
-	args = this.nopter.input(args);
-	
-	//if (testing===true && showArgs===true) return args;
-	
-	if (args.help === true)
-	{
-		log( this.nopter.help() );
-	}
-	else if (args.version === true)
-	{
-		log(pkg.version);
-	}
-	else if (args.input != null)
-	{
-		// TODO :: remove this when nopter's default values for Arrays are fixed
-		if (args.excludedKeywords === undefined)
-		{
-			args.excludedKeywords = defaultOptions.excludedKeywords;
+			opts[ renames[argName] ] = argValue;
 		}
-		
-		run(args.input,
+		else if (argName in DEFAULT_OPTIONS)
 		{
-			excludedKeywords:       args.excludedKeywords,
-			excludeExternalLinks:   args.excludeExternalLinks===true,
-			excludeInternalLinks:   args.excludeInternalLinks===true,
-			excludeLinksToSamePage: args.verbose!==true,
-			filterLevel:            args.filterLevel,
-			honorRobotExclusions:   args.followRobotExclusions!==true,
-			maxSockets:             args.maxSockets,
-			maxSocketsPerHost:      args.maxSocketsPerHost,
-			requestMethod:          args.get!==true ? "head" : "get",
-			userAgent:              args.userAgent
-		},
+			opts[argName] = argValue;
+		}
+		else if (args.get)
 		{
-			excludeCachedLinks:   args.verbose!==true,
-			excludeFilteredLinks: args.verbose!==true,
-			maintainLinkOrder:    args.maintainLinkOrder,
-			recursive:            args.recursive
-		});
+			opts.requestMethod = GET_METHOD;
+		}
+
+		return opts;
+	}, {});
+};
+
+
+
+const log = (...args) =>
+{
+	// Avoid spinner/progress chars getting stuck in the log
+	gauge.hide();
+
+	console.log(...args);
+
+	gauge.show();
+};
+
+
+
+log.page = pageURL =>
+{
+	log( white("\nGetting links from: ") + yellow(pageURL) );
+};
+
+
+
+log.page.metrics = () =>
+{
+	let output = gray(`Finished! ${stats.page.totalLinks} links found.`);
+
+	if (stats.page.skippedLinks > 0)
+	{
+		output += gray(` ${stats.page.skippedLinks} skipped.`);
+	}
+
+	if (stats.page.totalLinks > 0)
+	{
+		output += gray(" ");
+
+		if (stats.page.brokenLinks > 0)
+		{
+			output += red(`${stats.page.brokenLinks} broken`);
+		}
+		else
+		{
+			output += green(`${stats.page.brokenLinks} broken`);
+		}
+
+		output += gray(".");
+	}
+
+	log(output);
+};
+
+
+
+log.progress = () =>
+{
+	const links = checker.numActiveLinks + checker.numQueuedLinks;
+
+	const pageCompletion = links>0 ? 1/links : 0;
+
+	if (logOptions.recursive)
+	{
+		gauge.show(`Links:${links} Pages:${checker.numPages} Sites:${checker.numSites}`, pageCompletion);
 	}
 	else
 	{
-		log( nopter.error.fatal("Input URL required", "Use --help for more options", "Error") );
-		process.exit(1);
+		gauge.show(`Links:${links} Pages:${checker.numPages}`, pageCompletion);
 	}
 };
 
 
 
-function log()
+log.result = /*(*/result/*, finalResult)*/ =>
 {
-	// Avoid spinner chars getting stuck in the log
-	spinner.clear();
-	
-	console.log.apply(null, arguments);
-}
-
-
-
-function logMetrics(brokenLinks, excludedLinks, totalLinks, duration, preBreak, exit)
-{
-	var output = preBreak===true ? "\n" : "";
-	
-	output += chalk.gray("Finished! "+totalLinks+" links found.");
-	
-	if (excludedLinks > 0)
+	if (result.displayed)
 	{
-		output += chalk.gray(" "+excludedLinks+" excluded.");
-	}
-	
-	if (totalLinks > 0)
-	{
-		output += chalk.gray(" ");
-		output += chalk[ brokenLinks>0 ? "red" : "green" ](brokenLinks+" broken");
-		output += chalk.gray(".");
-	}
-	
-	if (duration != null)
-	{
-		output += chalk.gray("\nElapsed time: ");
-		output += chalk.gray( humanizeDuration(duration, {round:true, largest:2}) );
-	}
-	
-	log(output);
+		// @todo if the last result is skipped, the last RENDERED result will not be "└─"
+		let output = gray( /*finalResult!==true ?*/ "├─" /*: "└─"*/ );
 
-	if (exit === true)
-	{
-		process.exit(brokenLinks===0 ? 0 : 1);
-	}
-}
+		const {link} = result;
 
-
-
-/*
-	Ensure that `logMetrics()` is called after `logResults_delayed()`.
-*/
-function logMetrics_delayed(brokenLinks, excludedLinks, totalLinks, duration, preBreak, exit)
-{
-	setImmediate( function()
-	{
-		logMetrics(brokenLinks, excludedLinks, totalLinks, duration, preBreak, exit);
-	});
-}
-
-
-
-function logPage(data, pageUrl)
-{
-	var output = "";
-	
-	if (++data.total.pages > 1) output += "\n";
-	
-	output += chalk.white("Getting links from: ") + chalk.yellow(pageUrl);
-	
-	log(output);
-}
-
-
-
-function logResult(result, finalResult)
-{
-	var output = "";
-	
-	if (result.__cli_excluded !== true)
-	{
-		// TODO :: if later results are skipped, the last RENDERED result will not be "└─"
-		output = chalk.gray( finalResult!==true ? "├─" : "└─" );
-		
-		if (result.broken === true)
+		if (link.get(IS_BROKEN))
 		{
-			output += chalk.red("BROKEN");
-			output += chalk.gray("─ ");
+			output += red("BROKEN");
+			output += gray("─ ");
 		}
-		else if (result.excluded === true)
+		else if (link.get(WAS_EXCLUDED))
 		{
-			output += chalk.gray("─SKIP── ");
+			output += gray("─SKIP── ");
 		}
 		else
 		{
-			output += chalk.gray("──");
-			output += chalk.green("OK");
-			output += chalk.gray("─── ");
+			output += gray("──");
+			output += green("OK");
+			output += gray("─── ");
 		}
-		
-		if (result.url.resolved != null)
+
+		// @todo is ORIGINAL_URL only for invalid links?
+		output += yellow( link.get(REBASED_URL) ?? link.get(ORIGINAL_URL) );
+
+		if (link.get(IS_BROKEN))
 		{
-			output += chalk.yellow( result.url.resolved );
+			output += gray(` (${link.get(BROKEN_REASON)})`);
 		}
-		else
+		else if (link.get(WAS_EXCLUDED))
 		{
-			// Excluded scheme
-			output += chalk.yellow( result.url.original );
-		}
-		
-		if (result.broken === true)
-		{
-			output += chalk.gray(" ("+ result.brokenReason +")");
-		}
-		else if (result.excluded === true)
-		{
-			output += chalk.gray(" ("+ result.excludedReason +")");
+			output += gray(` (${link.get(EXCLUDED_REASON)})`);
 		}
 		// Don't display cached message if broken/excluded message is displayed
-		else if (result.http.cached === true)
+		else if (link.get(HTTP_RESPONSE_WAS_CACHED))
 		{
-			output += chalk.gray(" (CACHED)");
+			output += gray(" (CACHED)");
 		}
+
+		log(output);
 	}
-	
-	return output;
-}
+};
 
 
 
-/*
-	Logs links in the order that they are found in their containing HTML
-	document, even if later links receive an earlier response.
-*/
-function logResults(data)
+/**
+ * Logs links in the order that they are found in their containing HTML
+ * document, even if later links receive an earlier response.
+ */
+log.results = () =>
 {
-	var done,output,result;
-	var nextIsReady = true;
-	
-	while (nextIsReady)
+	// eslint-disable-next-line no-constant-condition
+	while (true)
 	{
-		result = data.page.results[data.page.currentIndex];
-		
+		const result = stats.page.results[stats.page.currentIndex];
+
 		if (result !== undefined)
 		{
-			done = data.page.done===true && data.page.currentIndex>=data.page.results.length-1;
-			
-			output = logResult(result, done);
-			
-			if (output !== "") log(output);
-			if (done === true) return;
-			
-			data.page.currentIndex++;
+			//const final = stats.page.currentIndex>=stats.page.results.length-1 && checker.numActiveLinks===0 && checker.numQueuedLinks===0;
+
+			log.result(result/*, final*/);
+
+			stats.page.currentIndex++;
 		}
 		else
 		{
-			nextIsReady = false;
+			break;
 		}
 	}
-}
+};
 
 
 
-/*
-	Ensure that `logResults()` is called after `data.page.done=true`.
-*/
-function logResults_delayed(data)
+log.site = () =>
 {
-	// Avoid more than one delay via multiple synchronous iterations
-	if (data.delay === null)
+	let output = "";
+
+	if (++stats.site.totalPages > 1)
 	{
-		data.delay = setImmediate( function()
-		{
-			logResults(data);
-			data.delay = null;
-		});
+		output += "\n";
 	}
-}
+
+	output += white("\nStarting recursive scan...");
+
+	log(output);
+};
 
 
 
-function pushResult(data, result, options)
+// @todo number of unique/uncached links
+// @todo "excluded links" [from cli] doesn't make sense with a value of 0 when there're skipped links in the log
+log.site.metrics = () =>
 {
-	if (options.maintainLinkOrder === true)
+	let output = "";
+	output += gray(`\nLinks found: ${stats.site.totalLinks}`);
+	output += gray(`\nLinks skipped: ${stats.site.skippedLinks}`);
+	output += gray(`\nLinks OK: ${stats.site.totalLinks - stats.site.skippedLinks - stats.site.brokenLinks}`);
+
+	let broken;
+
+	if (stats.site.totalLinks > 0)
 	{
-		data.page.results[result.html.index] = result;
+		broken = stats.site.brokenLinks>0 ? red : green;
 	}
 	else
 	{
-		data.page.results.push(result);
+		broken = gray;
 	}
-}
+
+	output += broken(`\nLinks broken: ${stats.site.brokenLinks}`);
+	output += gray("\nTime elapsed: ");
+	output += gray( humanizeDuration(Date.now() - stats.site.startTime, {largest:2, round:true}) );
+
+	const separator = gray("=".repeat( longest(stripAnsi(output).split("\n")).length ));
+
+	log(`\n${separator}${output}\n${separator}\n`);
+};
 
 
 
-function resetPageData(data)
+const run = () =>
 {
-	data.page.brokenLinks = 0;
-	data.page.currentIndex = 0;
-	data.page.done = false;
-	data.page.excludedLinks = 0;
-	data.page.results = [];
-	//data.page.startTime = Date.now();
-	data.page.totalLinks = 0;
-}
-
-
-
-function run(url, checkerOptions, logOptions)
-{
-	var handlers,instance;
-	var data = 
+	Object.values(gaugeThemes).forEach(theme =>
 	{
-		delay: null,
-		page: {},
-		total:
-		{
-			brokenLinks: 0,
-			excludedLinks: 0,
-			links: 0,
-			pages: 0,
-			startTime: Date.now()
-		}
-	};
-	
-	// In case first page doesn't call "html" handler
-	resetPageData(data);
-	
-	handlers =  
+		//theme.preProgressbar = `\n\n${theme.preProgressbar}`;
+		theme.preSubsection = gray("—");
+	});
+
+	gauge = new Gauge();
+	stats = new Statistics();
+
+	if (logOptions.recursive)
 	{
-		html: function(tree, robots, response, pageUrl)
+		checker = new SiteChecker(checkerOptions);
+	}
+	else
+	{
+		checker = new HtmlUrlChecker(checkerOptions);
+	}
+
+	checker
+	.on(HTML_EVENT, (tree, robots, response, pageURL) =>
+	{
+		log.page(pageURL);
+	})
+	.on(QUEUE_EVENT, () =>
+	{
+		log.progress();
+	})
+	.on(JUNK_EVENT, link =>
+	{
+		stats.pushResult(link);
+		log.progress();
+		log.results();
+	})
+	.on(LINK_EVENT, link =>
+	{
+		stats.pushResult(link);
+		log.progress();
+		log.results();
+	})
+	.on(PAGE_EVENT, (error, pageURL) =>
+	{
+		if (error != null)
 		{
-			resetPageData(data);
-			
-			logPage(data, pageUrl);
-		},
-		junk: function(result)
-		{
-			if (logOptions.excludeFilteredLinks === true)
+			// HTML_EVENT will not have been called
+			log.page(pageURL);
+
+			if (error.code<200 || error.code>299)
 			{
-				result.__cli_excluded = true;
-				
-				data.page.excludedLinks++;
-				data.total.excludedLinks++;
-			}
-			
-			data.page.totalLinks++;
-			data.total.links++;
-			
-			pushResult(data, result, logOptions);
-			
-			logResults_delayed(data);
-		},
-		link: function(result)
-		{
-			// Exclude cached links only if not broken
-			if (result.broken===false && result.http.cached===true && logOptions.excludeCachedLinks===true)
-			{
-				result.__cli_excluded = true;
-				
-				data.page.excludedLinks++;
-				data.total.excludedLinks++;
-			}
-			else if (result.broken === true)
-			{
-				data.page.brokenLinks++;
-				data.total.brokenLinks++;
-			}
-			
-			data.page.totalLinks++;
-			data.total.links++;
-			
-			pushResult(data, result, logOptions);
-			
-			logResults_delayed(data);
-		},
-		page: function(error, pageUrl)
-		{
-			if (error != null)
-			{
-				// "html" handler will not have been called
-				logPage(data, pageUrl);
-				
-				log( chalk[ error.code!==200 ? "red" : "gray" ](error.name+": "+error.message) );
+				log( red(`${error.name}: ${error.message}`) );
 			}
 			else
 			{
-				data.page.done = true;
-				
-				logMetrics_delayed(data.page.brokenLinks, data.page.excludedLinks, data.page.totalLinks);
+				log( gray(`${error.name}: ${error.message}`) );
 			}
-		},
-		end: function()
+
+			process.exitCode = 1;
+		}
+		// If more than a total of one page will be scanned
+		else if (logOptions.recursive || urls.length>1)
 		{
-			if (data.total.pages <= 0)
+			log.page.metrics();
+			log.progress();
+			stats.resetPage();
+
+			// If nothing after current page
+			if (checker.numPages === 1)
 			{
-				process.exit(1);
-			}
-			else if (data.total.pages === 1)
-			{
-				process.exit(data.page.done===true && data.total.brokenLinks===0 ? 0 : 1);
-			}
-			else if (data.total.pages > 1)
-			{
-				logMetrics_delayed(data.total.brokenLinks, data.total.excludedLinks, data.total.links, Date.now()-data.total.startTime, true, true);
+				log.site.metrics();
 			}
 		}
-	};
-	
-	if (logOptions.recursive !== true)
+		else
+		{
+			log.site.metrics();
+		}
+	})
+	.on(SITE_EVENT, () =>
 	{
-		instance = new blc.HtmlUrlChecker(checkerOptions, handlers);
+		log.site.metrics();
+		stats.resetSite();
+	})
+	.on(END_EVENT, () =>
+	{
+		// @todo store multiple site stats in an array and log all site metrics at very end?
+
+		if (supportsSemigraphics())
+		{
+			// Exit gracefully
+			clearTimeout(spinner);
+			gauge.disable();
+			keyScanner.release();
+
+			// @todo https://github.com/mikaelbr/node-notifier/issues/174
+			notifier.notify({ message:"Finished!", title });
+		}
+	})
+	.on(ERROR_EVENT, error =>
+	{
+		console.error(error);
+
+		// eslint-disable-next-line no-process-exit
+		process.exit(1);
+	});
+
+	if (logOptions.recursive)
+	{
+		log.site();
+	}
+
+	if (supportsSemigraphics())
+	{
+		// Show pause message
+		togglePause(false);
+
+		keyScanner = scanKeys(key =>
+		{
+			if (key.parsed === "space")
+			{
+				togglePause();
+			}
+		});
 	}
 	else
 	{
-		instance = new blc.SiteChecker(checkerOptions, handlers);
+		gauge.disable();
 	}
-	
-	spinner();
-	
-	instance.enqueue(url);
+
+	try
+	{
+		checker.pause();  // avoid auto-start
+		urls.forEach(url => checker.enqueue(new URL(url)));
+		checker.resume();  // start, if above didn't throw
+	}
+	catch ({message})
+	{
+		console.error(message);
+		process.exitCode = 1;
+	}
+};
+
+
+
+const spinnerInterval = () =>
+{
+	spinner = setTimeout(() =>
+	{
+		gauge.pulse(pauseMessage);
+		spinnerInterval();
+	}, 50);
+};
+
+
+
+class Statistics
+{
+	constructor()
+	{
+		this.page = {};
+		this.site = {};
+		this.resetSite();
+	}
+
+	pushResult(link)
+	{
+		const result = { displayed:true, link };
+
+		const hideCachedLink   = logOptions.hideCachedLinks   && link.get(IS_BROKEN)===false && link.get(HTTP_RESPONSE_WAS_CACHED);
+		const hideSkippedLink  = logOptions.hideSkippedLinks  && link.get(WAS_EXCLUDED);
+		const hideUnbrokenLink = logOptions.hideUnbrokenLinks && link.get(IS_BROKEN)===false;
+
+		if (hideCachedLink || hideSkippedLink || hideUnbrokenLink)
+		{
+			this.page.hiddenLinks++;
+			this.site.hiddenLinks++;
+			result.displayed = false;
+		}
+
+		if (link.get(IS_BROKEN))
+		{
+			this.page.brokenLinks++;
+			this.site.brokenLinks++;
+			process.exitCode = 1;
+		}
+		else if (link.get(WAS_EXCLUDED))
+		{
+			this.page.skippedLinks++;
+			this.site.skippedLinks++;
+		}
+
+		this.page.totalLinks++;
+		this.site.totalLinks++;
+
+		if (logOptions.maintainLinkOrder)
+		{
+			this.page.results[link.get(HTML_INDEX)] = result;
+		}
+		else
+		{
+			this.page.results.push(result);
+		}
+	}
+
+	resetPage()
+	{
+		this.page.brokenLinks = 0;
+		this.page.currentIndex = 0;
+		this.page.hiddenLinks = 0;
+		this.page.results = [];
+		this.page.skippedLinks = 0;
+		//this.page.startTime = Date.now();
+		this.page.totalLinks = 0;
+	}
+
+	resetSite()
+	{
+		this.resetPage();
+		this.site.brokenLinks = 0;
+		this.site.hiddenLinks = 0;
+		this.site.skippedLinks = 0;
+		this.site.startTime = Date.now();
+		this.site.totalLinks = 0;
+		this.site.totalPages = 0;
+	}
 }
 
 
 
-module.exports = cli;
+const togglePause = pause =>
+{
+	if (pause === undefined)
+	{
+		pause = !checker.isPaused;
+	}
+
+	if (pause)
+	{
+		checker.pause();
+
+		pauseMessage = `${yellow("PAUSED")}${gray(" — press space to resume")}`;
+		gauge.pulse(pauseMessage);
+		clearTimeout(spinner);
+	}
+	else
+	{
+		checker.resume();
+
+		pauseMessage = gray("press space to pause");
+		spinner = spinnerInterval();
+	}
+
+	log.progress();
+};
+
+
+
+export default (args=process.argv) =>
+{
+	const filterLevel =
+	[
+		"--filter-level:",
+		"  0: clickable links",
+		"  1: 0 + media, frames, meta refreshes",
+		"  2: 1 + stylesheets, scripts, forms",
+		"  3: 2 + metadata"
+	].join("\n");
+
+	const verbosity =
+	[
+		"--verbosity:",
+		"  0: broken links",
+		"  1: 0 + unbroken links",
+		"  2: 1 + skipped links"
+	].join("\n");
+
+	/* eslint-disable sort-keys */
+	const optionator = require("optionator")(
+	{
+		prepend: `${yellow(title.toUpperCase())}\n\n${green("Usage:")} blc [options] url1 [url2 ...]`,
+		append: `${gray(filterLevel)}\n\n${gray(verbosity)}\n`,
+		options:
+		[
+			{ heading:"Common Options" },
+			{ option:"recursive",        alias:"r", type:"Boolean",  description:`Recursively scan ("crawl") the HTML document(s)`,                   default:"false" },
+
+			{ heading:"Filtering Options" },
+			{ option:"exclude",                     type:"[String]", description:"Skip checking of links that match keywords/glob" },
+			{ option:"exclude-external", alias:"e", type:"Boolean",  description:"Skip checking of external links",                                   default:"false" },
+			{ option:"exclude-internal", alias:"i", type:"Boolean",  description:"Skip checking of internal links",                                   default:"false" },
+			{ option:"filter-level",                type:"Number",   description:"Include checking of links by HTML properties",                      default:`${DEFAULT_OPTIONS.filterLevel}` },
+			{ option:"follow",           alias:"f", type:"Boolean",  description:"Force-follow robot exclusions",                                     default:"false" },
+			{ option:"include",                     type:"[String]", description:"Only check links that match keywords/glob" },
+
+			{ heading:"Display Options" },
+			{ option:"help",             alias:"h", type:"Boolean",  description:"Display this help text",                                            default:"false" },
+			{ option:"ordered",          alias:"o", type:"Boolean",  description:"Maintain the order of links as they appear in their HTML document", default:"false" },
+			{ option:"verbosity",                   type:"Number",   description:"The display verbosity level",                                       default:"1" },
+			{ option:"version",          alias:"v", type:"Boolean",  description:"Display the app version",                                           default:"false" },
+
+			{ heading:"Advanced Options" },
+			{ option:"get",              alias:"g", type:"Boolean",  description:"Change request method to GET",                                      default:"false" },
+			{ option:"host-requests",               type:"Number",   description:"Concurrent requests limit per host",                                default:`${DEFAULT_OPTIONS.maxSocketsPerHost}` },
+			{ option:"requests",                    type:"Number",   description:"Concurrent requests limit ",                                        default:`${DEFAULT_OPTIONS.maxSockets}` },
+			{ option:"user-agent",                  type:"String",   description:"The user agent to use for checking links" }
+		]
+	});
+	/* eslint-disable sort-keys */
+
+	try
+	{
+		if (args === process.argv)
+		{
+			args = optionator.parseArgv(args);
+		}
+		else
+		{
+			args = optionator.parse(args);
+		}
+	}
+	catch (error)
+	{
+		args = error;
+	}
+
+	if (args instanceof Error)
+	{
+		console.error(args.message);
+		process.exitCode = 1;
+	}
+	else if (args.help)
+	{
+		console.log( optionator.generateHelp() );
+	}
+	else if (args.version)
+	{
+		console.log(packageVersion);
+	}
+	else if (args._.length > 0)
+	{
+		urls = args._;
+		checkerOptions = argsToOptions(args);
+		logOptions =
+		{
+			hideCachedLinks:   args.verbosity < 2,
+			hideSkippedLinks:  args.verbosity < 2,
+			hideUnbrokenLinks: args.verbosity < 1,
+			maintainLinkOrder: args.ordered,
+			recursive:         args.recursive === true  // default value is undefined
+		};
+
+		run();
+	}
+	else
+	{
+		console.error("At least one URL is required - see '--help'");
+		process.exitCode = 1;
+	}
+};
