@@ -1,23 +1,16 @@
-"use strict";
-var reasons = require("./internal/messages").reasons;
+import * as reasons from "./internal/reasons";
+import DEFAULT_OPTIONS from "./internal/defaultOptions";
+import HtmlChecker from "./public/HtmlChecker";
+import HtmlUrlChecker from "./public/HtmlUrlChecker";
+import SiteChecker from "./public/SiteChecker";
+import UrlChecker from "./public/UrlChecker";
 
+export * from "./internal/events";
+export * from "./internal/methods";
 
+// @todo https://github.com/tc39/proposal-export-default-from
+export {DEFAULT_OPTIONS};
+export {HtmlChecker, HtmlUrlChecker, SiteChecker, UrlChecker};
 
-var blc = 
-{
-	HtmlChecker:    require("./public/HtmlChecker"),
-	HtmlUrlChecker: require("./public/HtmlUrlChecker"),
-	SiteChecker:    require("./public/SiteChecker"),
-	UrlChecker:     require("./public/UrlChecker")
-};
-
-
-
-for (var i in reasons)
-{
-	blc[i] = reasons[i];
-}
-
-
-
-module.exports = blc;
+// @todo https://github.com/tc39/proposal-export-ns-from
+export {reasons};
