@@ -1,29 +1,32 @@
-"use strict";
-var pkg = require("../../package.json");
+import deepFreeze from "deep-freeze-node";
+import {HEAD_METHOD} from "./methods";
+import {name as packageName, version as packageVersion} from "../../package.json";
+import TAGS from "./tags";
+import userAgent from "default-user-agent";
 
-var userAgent = require("default-user-agent");
 
-var defaultOptions = 
+
+export default deepFreeze(
 {
-	acceptedSchemes: ["http","https"],
-	cacheExpiryTime: 3600000,
+	acceptedSchemes: ["http:","https:"],  // @todo add "file:"
+	cacheMaxAge: 3_600_000,
 	cacheResponses: true,
 	excludedKeywords: [],
-	excludedSchemes: ["data","geo","javascript","mailto","sms","tel"],
+	excludedSchemes: ["data:","geo:","javascript:","mailto:","sms:","tel:"],
 	excludeExternalLinks: false,
 	excludeInternalLinks: false,
-	excludeLinksToSamePage: true,
+	excludeLinksToSamePage: false,
 	filterLevel: 1,
 	honorRobotExclusions: true,
-	maxSockets: Infinity,
-	maxSocketsPerHost: 1,
+	includedKeywords: [],
+	includeLink: () => true,
+	includePage: () => true,
+	maxSockets: Infinity,  // @todo change to `maxExternalSockets`
+	maxSocketsPerHost: 1,  // @todo separate to `maxInternalSockets=5` and `maxExternalSocketsPerHost=1`
 	rateLimit: 0,
-	requestMethod: "head",
-	retry405Head: true,
-	tags: require("./tags"),
-	userAgent: userAgent(pkg.name, pkg.version)
-};
-
-
-
-module.exports = defaultOptions;
+	requestMethod: HEAD_METHOD,
+	retryHeadCodes: [405],
+	retryHeadFail: true,
+	tags: TAGS,
+	userAgent: userAgent(packageName, packageVersion)
+});
