@@ -1,33 +1,37 @@
-"use strict";
-var guard = require("robots-txt-guard");
-var parse = require("robots-txt-parse");
-
-var bhttp = require("bhttp");
-var urllib = require("url");
-var urlobj = require("urlobj");
-
+import {BLC_INVALID} from "./reasons";
+import {GET_METHOD} from "./methods";
+import guard from "robots-txt-guard";
+import isURL from "isurl";
+import parse from "robots-txt-parse";
+import requestHTTP from "./requestHTTP";
 
 
-function getRobotsTxt(url, options)
+
+/**
+ * Download and parse a robots.txt file from a server's root path.
+ * @param {URL} url
+ * @param {object} auth
+ * @param {URLCache} cache
+ * @param {object} options
+ * @throws {TypeError} non-URL
+ * @returns {Promise<guard>}
+ */
+export default async (url, auth, cache, options) =>
 {
-	url = urlobj.parse(url);
-	
-	// TODO :: this mutates the original (if was an object)
-	url.hash = null;
-	url.path = url.pathname = "/robots.txt";
-	url.query = null;
-	url.search = null;
-	
-	return bhttp.get(urllib.format(url),  // TODO :: https://github.com/joepie91/node-bhttp/issues/3
+	if (!isURL.lenient(url))
 	{
-		discardResponse: true,
-		headers: { "user-agent":options.userAgent },
-		stream: true
-	})
-	.then(parse)
-	.then(guard);
-}
+		throw new TypeError(BLC_INVALID);
+	}
+	else
+	{
+		url = new URL(url);
+		url.hash = "";
+		url.pathname = "/robots.txt";
+		url.search = "";
 
+		const {stream} = await requestHTTP(url, auth, GET_METHOD, cache, options);
 
-
-module.exports = getRobotsTxt;
+		// @todo https://github.com/tc39/proposal-pipeline-operator
+		return guard(await parse(stream));
+	}
+};
