@@ -1,57 +1,47 @@
-"use strict";
-var defaultOptions = require("./defaultOptions");
+import DEFAULT_OPTIONS from "./defaultOptions";
 
 
 
-/*
-	Convert an Array to a boolean-value Map.
-	
-	["asdf1","asdf2"]
-	
-	to
-	
-	{ asdf1:true, asdf2:true }
-*/
-function array2booleanMap(array)
+const HAS_BEEN_PARSED_VALUE = Symbol();
+
+
+
+/**
+ * Convert an Array to a map (object) of keys.
+ * @param {Array} array
+ * @returns {object}
+ * @example
+ * memoize(["asdf1", "asdf2"]);
+ * //-> { asdf1, asdf2 }
+ */
+const memoizeArray = array => array.reduce((map, value) =>
 {
-	var i,map,numElements;
-	
-	if (Array.isArray(array) === true)
-	{
-		map = {};
-		numElements = array.length;
-		
-		for (i=0; i<numElements; i++)
-		{
-			map[ array[i] ] = true;
-		}
-		
-		return map;
-	}
-	
-	// Unknown input -- return
-	return array;
-}
+	map[ value.toLowerCase() ] = true;
+	return map;
+}, {});
 
 
 
-function parseOptions(options)
+/**
+ * Combine consumer options with defaults, then normalize/optimize.
+ * @param {object} [options]
+ * @returns {object}
+ */
+export default (options = {}) =>
 {
-	if (options==null || options.__parsed!==true)
+	if (options.__parsed !== HAS_BEEN_PARSED_VALUE)
 	{
-		options = Object.assign({}, defaultOptions, options);
-		
-		// Maps have better search performance, but are not friendly for options
-		options.acceptedSchemes = array2booleanMap(options.acceptedSchemes);
-		options.excludedSchemes = array2booleanMap(options.excludedSchemes);
-		
-		// Undocumented -- avoids reparsing pass-thru options from class to class
-		options.__parsed = true;
+		options = { ...DEFAULT_OPTIONS, ...options };
+
+		// Maps of this kind are easier to work with, but are not consumer-friendly
+		options.acceptedSchemes = memoizeArray(options.acceptedSchemes);
+		options.excludedSchemes = memoizeArray(options.excludedSchemes);
+
+		options.requestMethod = options.requestMethod.toLowerCase();
+
+		// Undocumented -- avoids reparsing options passed through from class to class
+		options.__parsed = HAS_BEEN_PARSED_VALUE;
 	}
-	
+
 	return options;
-}
-
-
-
-module.exports = parseOptions;
+};
