@@ -1,134 +1,303 @@
-"use strict";
-var linkObj      = require("../internal/linkObj");
-var matchUrl     = require("../internal/matchUrl");
-var parseHtml    = require("../internal/parseHtml");
-var parseOptions = require("../internal/parseOptions");
-var scrapeHtml   = require("../internal/scrapeHtml");
+import * as reasons from "../internal/reasons";
+import {COMPLETE_EVENT, END_EVENT, ERROR_EVENT, HTML_EVENT, JUNK_EVENT, LINK_EVENT, QUEUE_EVENT} from "../internal/events";
+import {HTML_ATTR_NAME, HTML_ATTRS, HTML_INDEX, HTML_OFFSET_INDEX, HTML_TAG_NAME, IS_INTERNAL, IS_SAME_PAGE, REBASED_URL} from "../internal/Link";
+import isString from "is-string";
+import {map as linkTypes} from "link-types";
+import matchURL from "../internal/matchURL";
+import parseHTML from "../internal/parseHTML";
+import parseOptions from "../internal/parseOptions";
+import RobotDirectives, {NOFOLLOW, NOIMAGEINDEX, NOINDEX} from "robot-directives";
+import SafeEventEmitter from "../internal/SafeEventEmitter";
+import scrapeHTML from "../internal/scrapeHTML";
+import transitiveAuth from "../internal/transitiveAuth";
+import UrlChecker from "./UrlChecker";
 
-var UrlChecker = require("./UrlChecker");
-
-var isString = require("is-string");
-var linkTypes = require("link-types").map;
-var maybeCallback = require("maybe-callback");
-var RobotDirectives = require("robot-directives");
 
 
-
-function HtmlChecker(options, handlers)
+export default class HtmlChecker extends SafeEventEmitter
 {
-	var thisObj = this;
-	
-	reset(this);
-	
-	this.handlers = handlers || {};
-	this.options = options = parseOptions(options);
-	
-	this.urlChecker = new UrlChecker(this.options,
+	#auth;
+	#excludedLinks;
+	#options;
+	#resolvePromise;
+	#robots;
+	#scanning;
+	#urlChecker;
+
+
+
+	constructor(options)
 	{
-		link: function(result)
-		{
-			maybeCallback(thisObj.handlers.link)(result);
-		},
-		end: function()
-		{
-			// If stream finished
-			if (thisObj.parsed === true)
-			{
-				complete(thisObj);
-			}
-		}
-	});
-}
+		super();
+		this.#options = parseOptions(options);
+		this.#reset();
 
-
-
-HtmlChecker.prototype.clearCache = function()
-{
-	return this.urlChecker.clearCache();
-};
-
-
-
-HtmlChecker.prototype.numActiveLinks = function()
-{
-	return this.urlChecker.numActiveLinks();
-};
-
-
-
-HtmlChecker.prototype.numQueuedLinks = function()
-{
-	return this.urlChecker.numQueuedLinks();
-};
-
-
-
-HtmlChecker.prototype.pause = function()
-{
-	return this.urlChecker.pause();
-};
-
-
-
-HtmlChecker.prototype.resume = function()
-{
-	return this.urlChecker.resume();
-};
-
-
-
-HtmlChecker.prototype.scan = function(html, baseUrl, robots)
-{
-	var tree;
-	var thisObj = this;
-	
-	if (this.active === false)
-	{
-		// Prevent user error with undocumented arugment
-		if (robots instanceof RobotDirectives === false)
-		{
-			robots = new RobotDirectives({ userAgent: this.options.userAgent });
-		}
-		
-		this.active = true;
-		this.baseUrl = baseUrl;
-		this.robots = robots;
-		
-		parseHtml(html).then( function(document)
-		{
-			tree = document;
-			return scrapeHtml(document, thisObj.robots);
-		})
-		.then( function(links)
-		{
-			maybeCallback(thisObj.handlers.html)(tree, thisObj.robots);
-			
-			for (var i=0, numLinks=links.length; i<numLinks; i++)
-			{
-				enqueueLink(links[i], thisObj);
-			}
-			
-			thisObj.parsed = true;
-			
-			// If no links found or all links already checked
-			if (thisObj.urlChecker.numActiveLinks()===0 && thisObj.urlChecker.numQueuedLinks()===0)
-			{
-				complete(thisObj);
-			}
-		});
-		
-		return true;
+		this.#urlChecker = new UrlChecker(this.#options)
+		.on(ERROR_EVENT, error => this.emit(ERROR_EVENT, error))
+		.on(QUEUE_EVENT, () => this.emit(QUEUE_EVENT))
+		.on(LINK_EVENT, result => this.emit(LINK_EVENT, result))
+		.on(END_EVENT, () => this.#complete());
 	}
-	
-	return false;
-};
 
 
 
-HtmlChecker.prototype.__getCache = function()
-{
-	return this.urlChecker.__getCache();
-};
+	clearCache()
+	{
+		this.#urlChecker.clearCache();
+		return this;
+	}
+
+
+
+	#complete()
+	{
+		const resolvePromise = this.#resolvePromise;
+
+		this.#reset();
+
+		this.emit(COMPLETE_EVENT);
+
+		resolvePromise();
+	}
+
+
+
+	/**
+	 * Determine whether a Link should be excluded from checks, and the reason for such.
+	 * @param {Link} link
+	 * @returns {string|undefined}
+	 */
+	#getExcludeReason(link)
+	{
+		const attrName = link.get(HTML_ATTR_NAME);
+		const attrs = link.get(HTML_ATTRS);
+		const {href, protocol} = link.get(REBASED_URL);
+		const isInternal = link.get(IS_INTERNAL);
+		const tagName = link.get(HTML_TAG_NAME);
+
+		const
+		{
+			excludedKeywords,
+			excludedSchemes,
+			excludeExternalLinks,
+			excludeInternalLinks,
+			excludeLinksToSamePage,
+			honorRobotExclusions,
+			includedKeywords,
+			includeLink
+		} = this.#options;
+
+		if (this.#isExcludedAttribute(attrName, [tagName, "*"]))
+		{
+			return "BLC_HTML";
+		}
+		else if (excludeExternalLinks && isInternal===false)
+		{
+			return "BLC_EXTERNAL";
+		}
+		else if (excludeInternalLinks && isInternal)
+		{
+			return "BLC_INTERNAL";
+		}
+		else if (excludeLinksToSamePage && link.get(IS_SAME_PAGE))
+		{
+			return "BLC_SAMEPAGE";
+		}
+		else if (protocol in excludedSchemes)
+		{
+			return "BLC_SCHEME";
+		}
+		else if (honorRobotExclusions && this.#robots.oneIs([ NOFOLLOW, NOINDEX ]))
+		{
+			return "BLC_ROBOTS";
+		}
+		else if (honorRobotExclusions && this.#robots.is(NOIMAGEINDEX) && isRobotAttr(tagName, attrName))
+		{
+			return "BLC_ROBOTS";
+		}
+		else if (honorRobotExclusions && attrs?.rel!=null && linkTypes(attrs.rel).nofollow)
+		{
+			return "BLC_ROBOTS";
+		}
+		else if (matchURL(href, excludedKeywords))
+		{
+			return "BLC_KEYWORD";
+		}
+		else if (includedKeywords.length>0 && !matchURL(href, includedKeywords))
+		{
+			return "BLC_KEYWORD";
+		}
+		else
+		{
+			const filterResult = includeLink(link);
+
+			// Undocumented support for strings (from `SiteChecker`)
+			if (isString(filterResult) && filterResult in reasons)
+			{
+				return filterResult;
+			}
+			else if (!filterResult)
+			{
+				return "BLC_CUSTOM";
+			}
+			else
+			{
+				// Not excluded
+			}
+		}
+	}
+
+
+
+	/**
+	 * Determine whether a Link's HTML element and attribute would cause it to be excluded from checks.
+	 * @param {string} attrName
+	 * @param {Array<string>} tagNames
+	 * @returns {boolean}
+	 */
+	#isExcludedAttribute(attrName, tagNames)
+	{
+		const tagGroups = this.#options.tags[this.#options.filterLevel];
+
+		return tagNames.every(tagName => !(tagName in tagGroups) || !(attrName in tagGroups[tagName]));
+	}
+
+
+
+	get isPaused()
+	{
+		return this.#urlChecker.isPaused;
+	}
+
+
+
+	/**
+	 * Enqueue a Link if it is valid and passes filters.
+	 * @param {Link} link
+	 */
+	#maybeEnqueueLink(link)
+	{
+		if (link.get(REBASED_URL) === null)
+		{
+			link.set(HTML_OFFSET_INDEX, link.get(HTML_INDEX) - this.#excludedLinks);
+			link.break("BLC_INVALID");
+			link.include();
+
+			this.emit(LINK_EVENT, link);
+		}
+		else
+		{
+			const excludedReason = this.#getExcludeReason(link);
+
+			if (excludedReason === undefined)
+			{
+				link.set(HTML_OFFSET_INDEX, link.get(HTML_INDEX) - this.#excludedLinks);
+				link.include();
+
+				this.#urlChecker.enqueue(link, null, this.#auth);
+			}
+			else
+			{
+				link.set(HTML_OFFSET_INDEX, this.#excludedLinks++);
+				link.exclude(excludedReason);
+
+				this.emit(JUNK_EVENT, link);
+			}
+		}
+	}
+
+
+
+	get numActiveLinks()
+	{
+		return this.#urlChecker.numActiveLinks;
+	}
+
+
+
+	get numQueuedLinks()
+	{
+		return this.#urlChecker.numQueuedLinks;
+	}
+
+
+
+	pause()
+	{
+		this.#urlChecker.pause();
+		return this;
+	}
+
+
+
+	#reset()
+	{
+		this.#auth = null;
+		this.#excludedLinks = 0;
+		this.#resolvePromise = null;
+		this.#robots = null;
+		this.#scanning = false;
+	}
+
+
+
+	resume()
+	{
+		this.#urlChecker.resume();
+		return this;
+	}
+
+
+
+	// `robots` and `auth` are undocumented and for internal use only
+	async scan(html, baseURL, robots, auth)
+	{
+		if (this.#scanning)
+		{
+			throw new Error("Scan already in progress");
+		}
+		else
+		{
+			// Prevent user error with missing undocumented arugment
+			if (!(robots instanceof RobotDirectives))
+			{
+				robots = new RobotDirectives({ userAgent: this.#options.userAgent });
+			}
+
+			const transitive = transitiveAuth(baseURL, auth);
+
+			baseURL = transitive.url;  // @todo remove hash (and store somewhere?)
+
+			this.#auth = transitive.auth;
+			this.#robots = robots;
+			this.#scanning = true;
+
+			const document = await parseHTML(html);
+			const links = scrapeHTML(document, baseURL, this.#robots);  // @todo add auth?
+
+			this.emit(HTML_EVENT, document, this.#robots);
+
+			links.forEach(link => this.#maybeEnqueueLink(link));
+
+			const resolveOnComplete = new Promise(resolve => this.#resolvePromise = resolve);
+
+			// If no links found or all links already checked
+			if (this.#urlChecker.numActiveLinks===0 && this.#urlChecker.numQueuedLinks===0)
+			{
+				this.#complete();
+			}
+
+			return resolveOnComplete;
+		}
+	}
+
+
+
+	get __cache()
+	{
+		return this.#urlChecker.__cache;
+	}
+}
 
 
 
@@ -136,129 +305,10 @@ HtmlChecker.prototype.__getCache = function()
 
 
 
-function complete(instance)
+const isRobotAttr = (tagName, attrName) =>
 {
-	reset(instance);
-	
-	maybeCallback(instance.handlers.complete)();
-}
-
-
-
-function enqueueLink(link, instance)
-{
-	var excludedReason;
-	
-	linkObj.resolve(link, instance.baseUrl, instance.options);
-	
-	excludedReason = excludeLink(link, instance);
-	
-	if (excludedReason !== false)
-	{
-		link.html.offsetIndex = instance.excludedLinks++;
-		link.excluded = true;
-		link.excludedReason = excludedReason;
-		
-		linkObj.clean(link);
-		
-		maybeCallback(instance.handlers.junk)(link);
-		
-		return;
-	}
-	
-	link.html.offsetIndex = link.html.index - instance.excludedLinks;
-	link.excluded = false;
-	
-	instance.linkEnqueued = instance.urlChecker.enqueue(link);
-	
-	// TODO :: is this redundant? maybe use `linkObj.invalidate()` in `excludeLink()` ?
-	if (instance.linkEnqueued instanceof Error)
-	{
-		link.broken = true;
-		// TODO :: update limited-request-queue to support path-only URLs
-		link.brokenReason = instance.linkEnqueued.message==="Invalid URI" ? "BLC_INVALID" : "BLC_UNKNOWN";
-		
-		linkObj.clean(link);
-		
-		maybeCallback(instance.handlers.link)(link);
-	}
-}
-
-
-
-function excludeLink(link, instance)
-{
-	var attrSupported,externalFilter;
-	var attrName = link.html.attrName;
-	var tagName = link.html.tagName;
-	var tagGroup = instance.options.tags[instance.options.filterLevel][tagName];
-	
-	if (tagGroup != null)
-	{
-		attrSupported = tagGroup[attrName];
-	}
-	
-	if (attrSupported !== true) return "BLC_HTML";
-	if (instance.options.excludeExternalLinks===true   && link.internal===false) return "BLC_EXTERNAL";
-	if (instance.options.excludeInternalLinks===true   && link.internal===true)  return "BLC_INTERNAL";
-	if (instance.options.excludeLinksToSamePage===true && link.samePage===true)  return "BLC_SAMEPAGE";
-	if (instance.options.excludedSchemes[link.url.parsed.extra.protocolTruncated] === true) return "BLC_SCHEME";
-	
-	if (instance.options.honorRobotExclusions === true)
-	{
-		if (instance.robots.oneIs([ RobotDirectives.NOFOLLOW, RobotDirectives.NOINDEX ]) === true)
-		{
-			return "BLC_ROBOTS";
-		}
-		
-		if (instance.robots.is(RobotDirectives.NOIMAGEINDEX) === true)
-		{
-			if (
-			    (tagName==="img"      && attrName==="src"   ) || 
-			    (tagName==="input"    && attrName==="src"   ) || 
-			    (tagName==="menuitem" && attrName==="icon"  ) || 
-			    (tagName==="video"    && attrName==="poster")
-			   )
-			{
-				return "BLC_ROBOTS";
-			}
-		}
-		
-		if (link.html.attrs!=null && link.html.attrs.rel!=null && linkTypes(link.html.attrs.rel).nofollow===true)
-		{
-			return "BLC_ROBOTS";
-		}
-	}
-	
-	if (matchUrl(link.url.resolved, instance.options.excludedKeywords) === true) return "BLC_KEYWORD";
-	
-	// Undocumented handler for custom constraints
-	externalFilter = maybeCallback(instance.handlers._filter)(link);
-	
-	if (isString(externalFilter) === true)
-	{
-		return externalFilter;
-	}
-	/*else if (externalFilter === false)
-	{
-		return "BLC_CUSTOM";
-	}*/
-	
-	return false;
-}
-
-
-
-function reset(instance)
-{
-	instance.active = false;
-	instance.baseUrl = undefined;
-	instance.excludedLinks = 0;
-	instance.linkEnqueued = null;
-	instance.parsed = false;
-	instance.robots = null;
-}
-
-
-
-module.exports = HtmlChecker;
+	return (tagName==="img"      && attrName==="src"   ) ||
+	       (tagName==="input"    && attrName==="src"   ) ||
+	       (tagName==="menuitem" && attrName==="icon"  ) ||
+	       (tagName==="video"    && attrName==="poster");
+};
