@@ -1,194 +1,223 @@
-"use strict";
-var parseOptions = require("../internal/parseOptions");
-var streamHtml   = require("../internal/streamHtml");
-
-var HtmlChecker = require("./HtmlChecker");
-
-var maybeCallback = require("maybe-callback");
-var RequestQueue = require("limited-request-queue");
-var RobotDirectives = require("robot-directives");
-
+import {COMPLETE_EVENT, END_EVENT, ERROR_EVENT, HTML_EVENT, JUNK_EVENT, LINK_EVENT, PAGE_EVENT, QUEUE_EVENT} from "../internal/events";
+import HtmlChecker from "./HtmlChecker";
+import parseOptions from "../internal/parseOptions";
+import RequestQueue, {ITEM_EVENT, END_EVENT as REQUEST_QUEUE_END_EVENT} from "limited-request-queue";
+import RobotDirectives from "robot-directives";
+import SafeEventEmitter from "../internal/SafeEventEmitter";
+import streamHTML from "../internal/streamHTML";
+import transitiveAuth from "../internal/transitiveAuth";
 
 
-function HtmlUrlChecker(options, handlers)
+
+export default class HtmlUrlChecker extends SafeEventEmitter
 {
-	var thisObj = this;
-	
-	reset(this);
-	
-	this.handlers = handlers || {};
-	this.options = options = parseOptions(options);
-	
-	this.htmlUrlQueue = new RequestQueue(
+	#currentAuth;
+	#currentCustomData;
+	#currentDone;
+	#currentPageURL;
+	#currentResponse;
+	#currentRobots;
+	#htmlChecker;
+	#htmlUrlQueue;
+	#options;
+
+
+
+	constructor(options)
 	{
-		maxSockets: 1,
-		rateLimit: this.options.rateLimit
-	},
-	{
-		item: function(input, done)
+		super();
+		this.#reset();
+
+		this.#options = parseOptions(options);
+
+		this.#htmlUrlQueue = new RequestQueue(
 		{
-			thisObj.currentCustomData = input.data.customData;
-			thisObj.currentDone = done;
-			thisObj.currentPageUrl = input.url;
-			
-			streamHtml(thisObj.currentPageUrl, thisObj.__getCache(), thisObj.options).then( function(result)
+			maxSockets: 1,
+			rateLimit: this.#options.rateLimit
+		})
+		.on(ITEM_EVENT, async (url, {auth, customData}, done) =>
+		{
+			this.#reset();
+
+			this.#currentAuth = auth;
+			this.#currentCustomData = customData;
+			this.#currentDone = done;
+			this.#currentPageURL = url;  // @todo remove hash ?
+
+			try
 			{
-				thisObj.currentResponse = result.response;
-				
-				thisObj.currentRobots = new RobotDirectives({ userAgent: thisObj.options.userAgent });
-				
-				robotHeaders(thisObj);
-				
+				const {response, stream} = await streamHTML(this.#currentPageURL, this.#currentAuth, this.__cache, this.#options);
+
+				this.#currentResponse = response;
+				this.#currentRobots = new RobotDirectives({ userAgent: this.#options.userAgent });
+
+				this.#appendRobotHeaders();
+
 				// Passes robots instance so that headers are included in robot exclusion checks
-				thisObj.htmlChecker.scan(result.stream, result.response.url, thisObj.currentRobots);
-			})
-			.catch( function(error)
+				// @todo does the `await` cause `completedPage` to be called twice (other's in COMPLETE_EVENT) if error occurs?
+				await this.#htmlChecker.scan(stream, response.url, this.#currentRobots, this.#currentAuth);
+			}
+			catch (error)
 			{
-				completedPage(thisObj, error);
-			});
-		},
-		end: function()
+				this.#completedPage(error);
+			}
+		})
+		.on(REQUEST_QUEUE_END_EVENT, () =>
 		{
 			// Clear references for garbage collection
-			reset(thisObj);
-			
-			maybeCallback(thisObj.handlers.end)();
+			this.#reset();
+
+			this.emit(END_EVENT);
+		});
+
+		this.#htmlChecker = new HtmlChecker(this.#options)
+		.on(ERROR_EVENT, error => this.emit(ERROR_EVENT, error))
+		.on(HTML_EVENT, (tree, robots) =>
+		{
+			this.emit(HTML_EVENT, tree, robots, this.#currentResponse, this.#currentPageURL, this.#currentCustomData);
+		})
+		.on(QUEUE_EVENT, () => this.emit(QUEUE_EVENT))
+		.on(JUNK_EVENT, result => this.emit(JUNK_EVENT, result, this.#currentCustomData))
+		.on(LINK_EVENT, result => this.emit(LINK_EVENT, result, this.#currentCustomData))
+		.on(COMPLETE_EVENT, () => this.#completedPage());
+	}
+
+
+
+	/**
+	 * Append any robot headers.
+	 */
+	#appendRobotHeaders()
+	{
+		const xRobotsTag = this.#currentResponse.headers["x-robots-tag"];
+
+		// @todo https://github.com/nodejs/node/issues/3591
+		if (xRobotsTag != null)
+		{
+			this.#currentRobots.header(xRobotsTag);
 		}
-	});
-	
-	this.htmlChecker = new HtmlChecker(this.options,
+	}
+
+
+
+	clearCache()
 	{
-		html: function(tree, robots)
-		{
-			maybeCallback(thisObj.handlers.html)(tree, robots, thisObj.currentResponse, thisObj.currentPageUrl, thisObj.currentCustomData);
-		},
-		_filter: function(result)
-		{
-			// Undocumented handler for excluding links via custom constraints
-			return maybeCallback(thisObj.handlers._filter)(result);
-		},
-		junk: function(result)
-		{
-			maybeCallback(thisObj.handlers.junk)(result, thisObj.currentCustomData);
-		},
-		link: function(result)
-		{
-			maybeCallback(thisObj.handlers.link)(result, thisObj.currentCustomData);
-		},
-		complete: function()
-		{
-			completedPage(thisObj, null);
-		}
-	});
-}
+		this.#htmlChecker.clearCache();
+		return this;
+	}
 
 
 
-HtmlUrlChecker.prototype.clearCache = function()
-{
-	return this.htmlChecker.clearCache();
-};
-
-
-
-HtmlUrlChecker.prototype.dequeue = function(id)
-{
-	return this.htmlUrlQueue.dequeue(id);
-};
-
-
-
-HtmlUrlChecker.prototype.enqueue = function(pageUrl, customData)
-{
-	return this.htmlUrlQueue.enqueue(
+	/**
+	 * Emit PAGE_EVENT and continue the queue.
+	 * @param {Error} [error]
+	 */
+	#completedPage(error = null)
 	{
-		url: pageUrl,
-		data: { customData:customData }
-	});
-};
+		// @todo emit page error instead?
+		// @todo include redirected url if there is one?
+		this.emit(PAGE_EVENT, error, this.#currentPageURL, this.#currentCustomData);
+
+		// Auto-starts next queue item, if any
+		// Emits REQUEST_QUEUE_END_EVENT, if not
+		this.#currentDone();
+	}
 
 
 
-HtmlUrlChecker.prototype.numActiveLinks = function()
-{
-	return this.htmlChecker.numActiveLinks();
-};
-
-
-
-HtmlUrlChecker.prototype.numPages = function()
-{
-	return this.htmlUrlQueue.length();
-};
-
-
-
-HtmlUrlChecker.prototype.numQueuedLinks = function()
-{
-	return this.htmlChecker.numQueuedLinks();
-};
-
-
-
-HtmlUrlChecker.prototype.pause = function()
-{
-	this.htmlChecker.pause();
-	return this.htmlUrlQueue.pause();
-};
-
-
-
-HtmlUrlChecker.prototype.resume = function()
-{
-	this.htmlChecker.resume();
-	return this.htmlUrlQueue.resume();
-};
-
-
-
-HtmlUrlChecker.prototype.__getCache = function()
-{
-	return this.htmlChecker.__getCache();
-};
-
-
-
-//::: PRIVATE FUNCTIONS
-
-
-
-function completedPage(instance, error)
-{
-	maybeCallback(instance.handlers.page)(error, instance.currentPageUrl, instance.currentCustomData);
-	
-	// Auto-starts next queue item, if any
-	// If not, fires "end"
-	instance.currentDone();
-}
-
-
-
-function reset(instance)
-{
-	instance.currentCustomData = null;
-	instance.currentDone = null;
-	instance.currentPageUrl = null;
-	instance.currentResponse = null;
-	instance.currentRobots = null;
-}
-
-
-
-function robotHeaders(instance)
-{
-	// TODO :: https://github.com/joepie91/node-bhttp/issues/20
-	// TODO :: https://github.com/nodejs/node/issues/3591
-	if (instance.currentResponse.headers["x-robots-tag"] != null)
+	dequeue(id)
 	{
-		instance.currentRobots.header( instance.currentResponse.headers["x-robots-tag"] );
+		const success = this.#htmlUrlQueue.dequeue(id);
+
+		this.emit(QUEUE_EVENT);
+
+		return success;
+	}
+
+
+
+	// `auth` is undocumented and for internal use only
+	enqueue(pageURL, customData, auth)
+	{
+		// @todo this could get messy if there're many different credentials per site (if we cache based on headers)
+		const transitive = transitiveAuth(pageURL, auth);
+
+		const id = this.#htmlUrlQueue.enqueue(transitive.url, { auth:transitive.auth, customData });
+
+		this.emit(QUEUE_EVENT);
+
+		return id;
+	}
+
+
+
+	has(id)
+	{
+		return this.#htmlUrlQueue.has(id);
+	}
+
+
+
+	get isPaused()
+	{
+		return this.#htmlChecker.isPaused;
+	}
+
+
+
+	get numActiveLinks()
+	{
+		return this.#htmlChecker.numActiveLinks;
+	}
+
+
+
+	get numPages()
+	{
+		return this.#htmlUrlQueue.length;
+	}
+
+
+
+	get numQueuedLinks()
+	{
+		return this.#htmlChecker.numQueuedLinks;
+	}
+
+
+
+	pause()
+	{
+		this.#htmlChecker.pause();
+		this.#htmlUrlQueue.pause();
+		return this;
+	}
+
+
+
+	#reset()
+	{
+		this.#currentAuth = null;
+		this.#currentCustomData = null;
+		this.#currentDone = null;
+		this.#currentPageURL = null;
+		this.#currentResponse = null;
+		this.#currentRobots = null;
+	}
+
+
+
+	resume()
+	{
+		this.#htmlChecker.resume();
+		this.#htmlUrlQueue.resume();
+		return this;
+	}
+
+
+
+	get __cache()
+	{
+		return this.#htmlChecker.__cache;
 	}
 }
-
-
-
-module.exports = HtmlUrlChecker;
