@@ -1,139 +1,144 @@
-"use strict";
-var checkUrl     = require("../internal/checkUrl");
-var linkObj      = require("../internal/linkObj");
-var parseOptions = require("../internal/parseOptions");
-
-var isString = require("is-string");
-var maybeCallback = require("maybe-callback");
-var RequestQueue = require("limited-request-queue");
-var UrlCache = require("urlcache");
-var urlobj = require("urlobj");
+import checkLink from "../internal/checkLink";
+import {END_EVENT, LINK_EVENT, QUEUE_EVENT} from "../internal/events";
+import isURL from "isurl";
+import Link, {REBASED_URL} from "../internal/Link";
+import parseOptions from "../internal/parseOptions";
+import RequestQueue, {ITEM_EVENT, END_EVENT as REQUEST_QUEUE_END_EVENT} from "limited-request-queue";
+import SafeEventEmitter from "../internal/SafeEventEmitter";
+import URLCache from "urlcache";
 
 
 
-function UrlChecker(options, handlers)
+export default class UrlChecker extends SafeEventEmitter
 {
-	var thisObj = this;
-	
-	this.handlers = handlers || {};
-	this.options = options = parseOptions(options);
+	#cache;
+	#linkQueue;
 
-	this.cache = new UrlCache(
+
+
+	constructor(options)
 	{
-		expiryTime: this.options.cacheExpiryTime,
-		normalizeUrls: false
-	});
-	
-	this.linkQueue = new RequestQueue(
-	{
-		maxSockets:        this.options.maxSockets,
-		maxSocketsPerHost: this.options.maxSocketsPerHost,
-		rateLimit:         this.options.rateLimit
-	},
-	{
-		item: function(input, done)
+		super();
+
+		options = parseOptions(options);
+
+		this.#cache = new URLCache({ maxAge:options.cacheMaxAge });
+
+		this.#linkQueue = new RequestQueue(
 		{
-			// TODO :: make this more reusable
-			function handle_checkUrl(result)
-			{
-				maybeCallback(thisObj.handlers.link)(result, input.data.customData);
-				
-				// Auto-starts next queue item, if any
-				// If not, fires "end"
-				done();
-			}
-			
-			if (input.data.linkObj !== undefined)
-			{
-				checkUrl(input.data.linkObj, null, thisObj.cache, thisObj.options).then(handle_checkUrl);
-			}
-			else
-			{
-				// TODO :: send url object -- remove orgUrl?
-				checkUrl(input.data.orgUrl, input.data.baseUrl, thisObj.cache, thisObj.options).then(handle_checkUrl);
-			}
-		},
-		end: function()
+			maxSockets:        options.maxSockets,
+			maxSocketsPerHost: options.maxSocketsPerHost,
+			rateLimit:         options.rateLimit
+		})
+		.on(ITEM_EVENT, async (url, {auth, customData, link}, done) =>
 		{
-			maybeCallback(thisObj.handlers.end)();
+			const result = await checkLink(link, auth, this.#cache, options);
+
+			this.emit(LINK_EVENT, result, customData);
+
+			// Auto-starts next queue item, if any
+			// Emits REQUEST_QUEUE_END_EVENT, if not
+			done();
+		})
+		.on(REQUEST_QUEUE_END_EVENT, () => this.emit(END_EVENT));
+	}
+
+
+
+	clearCache()
+	{
+		this.#cache.clear();
+		return this;
+	}
+
+
+
+	dequeue(id)
+	{
+		const success = this.#linkQueue.dequeue(id);
+
+		this.emit(QUEUE_EVENT);
+
+		return success;
+	}
+
+
+
+	// `auth` is undocumented and for internal use only
+	enqueue(url, customData, auth={})
+	{
+		let link;
+
+		// Undocumented internal use: `enqueue(Link)`
+		if (url instanceof Link)
+		{
+			link = url;
 		}
-	});
+		// Documented use: `enqueue(URL)`
+		else if (isURL.lenient(url))
+		{
+			link = new Link().resolve(url);
+		}
+		else
+		{
+			throw new TypeError("Invalid URL");
+		}
+
+		const id = this.#linkQueue.enqueue(link.get(REBASED_URL), { auth, customData, link });
+
+		this.emit(QUEUE_EVENT);
+
+		return id;
+	}
+
+
+
+	has(id)
+	{
+		return this.#linkQueue.has(id);
+	}
+
+
+
+	get isPaused()
+	{
+		return this.#linkQueue.isPaused;
+	}
+
+
+
+	get numActiveLinks()
+	{
+		return this.#linkQueue.numActive;
+	}
+
+
+
+	get numQueuedLinks()
+	{
+		return this.#linkQueue.numQueued;
+	}
+
+
+
+	pause()
+	{
+		this.#linkQueue.pause();
+		return this;
+	}
+
+
+
+	resume()
+	{
+		this.#linkQueue.resume();
+		return this;
+	}
+
+
+
+	get __cache()
+	{
+		return this.#cache;
+	}
 }
-
-
-
-UrlChecker.prototype.clearCache = function()
-{
-	return this.cache.clear();
-};
-
-
-
-UrlChecker.prototype.dequeue = function(id)
-{
-	return this.linkQueue.dequeue(id);
-};
-
-
-
-UrlChecker.prototype.enqueue = function(url, baseUrl, customData)
-{
-	// Undocumented internal use: enqueue(linkObj)
-	if (isString(url)===false && url.broken_link_checker===true)
-	{
-		return this.linkQueue.enqueue(
-		{
-			url: url.url.parsed,
-			data: { customData:customData, linkObj:url }
-		});
-	}
-	// Documented use: enqueue(url, baseUrl)
-	// or erroneous and let linkQueue sort it out
-	else
-	{
-		return this.linkQueue.enqueue(
-		{
-			url: urlobj.resolve(baseUrl || "", urlobj.parse(url) ),  // URL must be absolute
-			data: { orgUrl:url, baseUrl:baseUrl, customData:customData }
-		});
-	}
-};
-
-
-
-UrlChecker.prototype.numActiveLinks = function()
-{
-	return this.linkQueue.numActive();
-};
-
-
-
-UrlChecker.prototype.numQueuedLinks = function()
-{
-	return this.linkQueue.numQueued();
-};
-
-
-
-UrlChecker.prototype.pause = function()
-{
-	return this.linkQueue.pause();
-};
-
-
-
-UrlChecker.prototype.resume = function()
-{
-	return this.linkQueue.resume();
-};
-
-
-
-UrlChecker.prototype.__getCache = function()
-{
-	return this.cache;
-};
-
-
-
-module.exports = UrlChecker;
