@@ -1,33 +1,64 @@
-"use strict";
-var server        = require("./server");
-var testGenerator = require("./testGenerator");
+import chai from "chai";
+import chaiAsPromised from "chai-as-promised";
+import chaiSubset from "chai-subset";
+import chaiThings from "chai-things";
+import tagsString from "./tagsString";
 
-var chai = require("chai");
-chai.config.includeStack = true;
-chai.use( require("chai-as-promised") );
-chai.use( require("chai-like") );
-chai.use( require("chai-things") );
+export * from "./fixture";
+export * from "./options";
 
-require("es6-promise").polyfill();
-require("object.assign").shim();
-
-
-
-module.exports = 
+export
 {
-	a_an:       testGenerator.a_an,
-	addSlashes: testGenerator.addSlashes,
-	format:     testGenerator.format,
-	//italic:     testGenerator.italic,
-	
-	options:    require("./options"),
-	
-	startConnection:  server.startConnection,
-	startConnections: server.startConnections,
-	stopConnection:   server.stopConnection,
-	stopConnections:  server.stopConnections,
-	
-	tagsString: require("./tagsString"),
-	
-	fixture: require("./fixture")
-};
+	start as startServer,
+	start as startServers,
+	startDead as startDeadServer,
+	startDead as startDeadServers,
+	stop as stopServers
+} from "./server";
+
+// @todo https://github.com/tc39/proposal-export-default-from
+export {tagsString};
+
+
+
+chai
+.use(chaiAsPromised)
+.use(chaiSubset)
+.use(chaiThings)
+.config.includeStack = true;
+
+
+
+export const defaultAuth = Object.freeze(
+{
+	password: "",
+	username: ""
+});
+
+
+
+export class ExpectedError extends Error
+{
+	constructor()
+	{
+		super("This was thrown so that it could be caught");
+	}
+}
+
+
+
+export const simplifyLink = link => link.toJSON();
+
+export const simplifyLinks = links => links.map(simplifyLink);
+
+export const simplifyPageLinks = pages => pages.map(simplifyLinks);
+
+
+
+export class WrongCallError extends Error
+{
+	constructor()
+	{
+		super("This should not have been called");
+	}
+}
