@@ -1,36 +1,31 @@
-"use strict";
-var appDefaultOptions = require("../../lib/internal/defaultOptions");
-var parseOptions      = require("../../lib/internal/parseOptions");
+import appDefaultOptions from "../../lib/internal/defaultOptions";
+import deepFreeze from "deep-freeze-node";
+import parseOptions from "../../lib/internal/parseOptions";
 
-var testDefaultOptions =
+
+
+export const parsedOptions = overrides => deepFreeze(parseOptions(testOptions(overrides)));
+
+export const rawOptions = overrides => deepFreeze(testOptions(overrides));
+
+
+
+const testDefaultOptions =
 {
 	// All other options will use default values
 	// as this will ensure that when they change, tests WILL break
 	cacheResponses: false,
-	excludeInternalLinks: false,
-	excludeLinksToSamePage: false,
 	filterLevel: 3,
 	honorRobotExclusions: false,
-	maxSockets: Infinity,
 	maxSocketsPerHost: Infinity,
-	retry405Head: false
+	retryHeadFail: false
 };
 
 
 
-function options(overrides)
-{
-	overrides = Object.assign
-	(
-		{},
-		appDefaultOptions,
-		testDefaultOptions,
-		overrides
-	);
-	
-	return parseOptions(overrides);
-}
-
-
-
-module.exports = options;
+const testOptions = overrides =>
+({
+	...appDefaultOptions,
+	...testDefaultOptions,
+	...overrides
+});
